feat(server): integrate node-fetch for external API calls, add Movie schema and resolvers
This commit introduces the use of `node-fetch` for making external API calls to the YTS movie database. It includes the addition of a new `Movie` schema in the GraphQL server setup, along with corresponding query resolvers for fetching all movies and individual movie details. The `package.json` and `package-lock.json` files are updated to include `node-fetch` as a dependency. Also, minor modifications are made to the server's code structure to accommodate these changes.